+++ package.json
@@ -23,9 +23,7 @@
         "pinia": "^2.2.4",
         "tiny-emitter": "^2.1.0",
         "typescript": "^5.6.3",
-        "vue": "^3.5.12",
-        "uuid": "^9.0.1",
-        "card-validator": "^8.1.1"
+        "vue": "^3.5.12"
     },
     "devDependencies": {
         "eslint": "^9.15.0",
