--- src/models/enums/compPoints.ts
@@ -1,20 +1,0 @@
-export const allType = "All";
-export enum CompPointsTypes {
-    FREE_SPINS = "FREE_SPINS",
-    MONEY_REWARD = "MONEY_REWARD",
-    SPECIAL_REWARDS = "SPECIAL_REWARDS",
-}
-
-export enum CompPointRatesTypes {
-    MONEY = "money",
-    LOOTBOXES = "lootboxes",
-    FREESPINS = "freespins",
-    LOTTERIES = "lotteries",
-}
-
-export enum CompPointTypes {
-    popular= "POPULAR",
-    new= "NEW"
-}
-
-export default {};
