--- src/models/enums/currencies.ts
@@ -1,24 +1,0 @@
-export enum Currencies {
-    EUR = "EUR",
-    USD = "USD",
-    CAD = "CAD",
-    AUD = "AUD",
-    NOK = "NOK",
-    PLN = "PLN",
-    NZD = "NZD",
-    JPY = "JPY",
-    BRL = "BRL",
-    INR = "INR",
-    BTC = "BTC",
-    BCH = "BCH",
-    ETH = "ETH",
-    LTC = "LTC",
-    DOG = "DOG",
-    USDT = "USDT",
-    USDC = "USDC",
-    BNB = "BNB",
-    ADA = "ADA",
-    TRX = "TRX",
-}
-
-export default {};
