--- src/models/enums/formFields.ts
@@ -1,17 +1,0 @@
-export enum FormFields {
-    first_name = "first_name",
-    last_name = "last_name",
-    nickname = "nickname",
-    date_of_birth = "date_of_birth",
-    gender = "gender",
-    country = "country",
-    city = "city",
-    address = "address",
-    postal_code = "postal_code",
-    receive_promos = "receive_promos",
-    receive_sms_promos = "receive_sms_promos",
-    mobile_phone = "mobile_phone",
-    state = "state",
-}
-
-export default {};
