--- src/models/utmMetrics.ts
@@ -1,7 +1,0 @@
-export interface IUtmMetrics {
-    utm_source: string;
-    utm_medium: string;
-    utm_campaign: string;
-    utm_content: string;
-    utm_term: string;
-}
