+++ src/services/api/DTO/cashbox.ts
@@ -1,4 +1,25 @@
-import { Currencies } from "../../../models/enums/currencies";
+export enum Currencies {
+    EUR = "EUR",
+    USD = "USD",
+    NOK = "NOK",
+    CAD = "CAD",
+    PLN = "PLN",
+    NZD = "NZD",
+    JPY = "JPY",
+    AUD = "AUD",
+    BRL = "BRL",
+    INR = "INR",
+    USDC = "USDC",
+    BTC = "BTC",
+    ETH = "ETH",
+    LTC = "LTC",
+    BCH = "BCH",
+    DOG = "DOG",
+    USDT = "USDT",
+    BNB = "BNB",
+    ADA = "ADA",
+    TRX = "TRX",
+}
 
 export enum ActionsTransaction {
     DEPOSIT = "deposit",
@@ -37,4 +58,3 @@ export interface IPlayerPayment {
     cancellation_reason?: string;
 
 }
-export default {};
