--- src/services/api/DTO/couponePromoCodes.ts
@@ -1,12 +1,0 @@
-export interface IActiveCouponResp {
-    "bonus_id": string;
-    "name": string;
-    "bonus_type": string;
-}
-
-export enum UserCouponStatuses {
-    alreadyActivated = "already_activated",
-    failedActivate = "failed_to_activate",
-    successfullyActivated = "successfully_activated",
-}
-export default {};
