--- src/services/api/DTO/multilang.ts
@@ -1,10 +1,0 @@
-export type LocaleName = string;
-
-export interface ILocale {
-    code: string;
-    name: LocaleName;
-    name_in_locale: string;
-    default: boolean;
-}
-
-export type Locales = ILocale[];
