--- src/services/api/DTO/performanceDTO.ts
@@ -1,9 +1,0 @@
-export interface IGameObserveDTO {
-    country: string;
-    userId: string;
-    gameId: string;
-    gameName: string;
-    provider: string;
-    value: number;
-    connectionType: string;
-}
