--- src/services/api/requests/couponePromoCodes.ts
@@ -1,13 +1,0 @@
-import log from "../../../controllers/Logger";
-import { IActiveCouponResp } from "../DTO/couponePromoCodes";
-import { http } from "../http";
-
-export async function activeCouponReq(code: string): Promise<IActiveCouponResp> {
-    try {
-        const { data } = await http().post<IActiveCouponResp>("/api/v2/bonuses/promo-code", { code });
-        return data;
-    } catch (err) {
-        log.error("ACTIVE_COUPON_ERROR", err);
-        throw err;
-    }
-}
