--- src/services/api/requests/mockLootbox/lootbox.ts
@@ -1,664 +1,0 @@
-export default [
-    {
-        weight: 4000,
-        prize: false,
-        created_at: "2024-11-07T13:46:44.064Z",
-        bonuses: [
-            {
-                title: "Rocket Wheel 10 FS",
-                type: "freespins",
-                attributes: [
-                    {
-                        field: "freespins_count",
-                        value: 10,
-                    },
-                ],
-                result_bonus: [
-                    {
-                        field: "wager",
-                        value: 30,
-                    },
-                    {
-                        field: "max_win",
-                        value: [
-                            {
-                                currency: "EUR",
-                                amount_cents: 100000,
-                            },
-                            {
-                                currency: "USD",
-                                amount_cents: 100000,
-                            },
-                            {
-                                currency: "NOK",
-                                amount_cents: 1050000,
-                            },
-                            {
-                                currency: "CAD",
-                                amount_cents: 150000,
-                            },
-                            {
-                                currency: "NZD",
-                                amount_cents: 150000,
-                            },
-                            {
-                                currency: "AUD",
-                                amount_cents: 150000,
-                            },
-                        ],
-                    },
-                ],
-            },
-        ],
-        id: 767669,
-    },
-    {
-        weight: 4000,
-        prize: false,
-        created_at: "2024-11-07T13:46:44.069Z",
-        bonuses: [
-            {
-                title: "Rocket Wheel 20 FS",
-                type: "freespins",
-                attributes: [
-                    {
-                        field: "freespins_count",
-                        value: 20,
-                    },
-                ],
-                result_bonus: [
-                    {
-                        field: "wager",
-                        value: 30,
-                    },
-                    {
-                        field: "max_win",
-                        value: [
-                            {
-                                currency: "EUR",
-                                amount_cents: 100000,
-                            },
-                            {
-                                currency: "USD",
-                                amount_cents: 100000,
-                            },
-                            {
-                                currency: "NOK",
-                                amount_cents: 1050000,
-                            },
-                            {
-                                currency: "CAD",
-                                amount_cents: 150000,
-                            },
-                            {
-                                currency: "NZD",
-                                amount_cents: 150000,
-                            },
-                            {
-                                currency: "AUD",
-                                amount_cents: 150000,
-                            },
-                        ],
-                    },
-                ],
-            },
-        ],
-        id: 767670,
-    },
-    {
-        weight: 2000,
-        prize: false,
-        created_at: "2024-11-07T13:46:44.074Z",
-        bonuses: [
-            {
-                title: "Rocket Wheel 50 FS",
-                type: "freespins",
-                attributes: [
-                    {
-                        field: "freespins_count",
-                        value: 50,
-                    },
-                ],
-                result_bonus: [
-                    {
-                        field: "wager",
-                        value: 30,
-                    },
-                    {
-                        field: "max_win",
-                        value: [
-                            {
-                                currency: "EUR",
-                                amount_cents: 100000,
-                            },
-                            {
-                                currency: "USD",
-                                amount_cents: 100000,
-                            },
-                            {
-                                currency: "NOK",
-                                amount_cents: 1050000,
-                            },
-                            {
-                                currency: "CAD",
-                                amount_cents: 150000,
-                            },
-                            {
-                                currency: "NZD",
-                                amount_cents: 150000,
-                            },
-                            {
-                                currency: "AUD",
-                                amount_cents: 150000,
-                            },
-                        ],
-                    },
-                ],
-            },
-        ],
-        id: 767671,
-    },
-    {
-        weight: 0,
-        prize: false,
-        created_at: "2024-11-07T13:46:44.078Z",
-        bonuses: [
-            {
-                title: "Rocket Wheel 100 FS",
-                type: "freespins",
-                attributes: [
-                    {
-                        field: "freespins_count",
-                        value: 100,
-                    },
-                ],
-                result_bonus: [
-                    {
-                        field: "wager",
-                        value: 30,
-                    },
-                    {
-                        field: "max_win",
-                        value: [
-                            {
-                                currency: "EUR",
-                                amount_cents: 100000,
-                            },
-                            {
-                                currency: "USD",
-                                amount_cents: 100000,
-                            },
-                            {
-                                currency: "NOK",
-                                amount_cents: 1050000,
-                            },
-                            {
-                                currency: "CAD",
-                                amount_cents: 150000,
-                            },
-                            {
-                                currency: "NZD",
-                                amount_cents: 150000,
-                            },
-                            {
-                                currency: "AUD",
-                                amount_cents: 150000,
-                            },
-                        ],
-                    },
-                ],
-            },
-        ],
-        id: 767672,
-    },
-    {
-        weight: 0,
-        prize: false,
-        created_at: "2024-11-07T13:46:44.083Z",
-        bonuses: [
-            {
-                title: "Rocket Wheel 150 FS",
-                type: "freespins",
-                attributes: [
-                    {
-                        field: "freespins_count",
-                        value: 150,
-                    },
-                ],
-                result_bonus: [
-                    {
-                        field: "wager",
-                        value: 30,
-                    },
-                    {
-                        field: "max_win",
-                        value: [
-                            {
-                                currency: "EUR",
-                                amount_cents: 100000,
-                            },
-                            {
-                                currency: "USD",
-                                amount_cents: 100000,
-                            },
-                            {
-                                currency: "NOK",
-                                amount_cents: 1050000,
-                            },
-                            {
-                                currency: "CAD",
-                                amount_cents: 150000,
-                            },
-                            {
-                                currency: "NZD",
-                                amount_cents: 150000,
-                            },
-                            {
-                                currency: "AUD",
-                                amount_cents: 150000,
-                            },
-                        ],
-                    },
-                ],
-            },
-        ],
-        id: 767673,
-    },
-    {
-        weight: 0,
-        prize: false,
-        created_at: "2024-11-07T13:46:44.089Z",
-        bonuses: [
-            {
-                title: "Rocket Wheel 250 FS",
-                type: "freespins",
-                attributes: [
-                    {
-                        field: "freespins_count",
-                        value: 250,
-                    },
-                ],
-                result_bonus: [
-                    {
-                        field: "wager",
-                        value: 30,
-                    },
-                    {
-                        field: "max_win",
-                        value: [
-                            {
-                                currency: "EUR",
-                                amount_cents: 100000,
-                            },
-                            {
-                                currency: "USD",
-                                amount_cents: 100000,
-                            },
-                            {
-                                currency: "NOK",
-                                amount_cents: 1050000,
-                            },
-                            {
-                                currency: "CAD",
-                                amount_cents: 150000,
-                            },
-                            {
-                                currency: "NZD",
-                                amount_cents: 150000,
-                            },
-                            {
-                                currency: "AUD",
-                                amount_cents: 150000,
-                            },
-                        ],
-                    },
-                ],
-            },
-        ],
-        id: 767674,
-    },
-    {
-        weight: 0,
-        prize: false,
-        created_at: "2024-11-07T13:46:44.093Z",
-        bonuses: [
-            {
-                title: "Rocket Wheel Bonus 1500 EUR",
-                type: "bonus",
-                attributes: [
-                    {
-                        field: "bonus_amount",
-                        value: [
-                            {
-                                currency: "EUR",
-                                amount_cents: 150000,
-                            },
-                            {
-                                currency: "USD",
-                                amount_cents: 150000,
-                            },
-                            {
-                                currency: "CAD",
-                                amount_cents: 230000,
-                            },
-                            {
-                                currency: "AUD",
-                                amount_cents: 230000,
-                            },
-                            {
-                                currency: "NOK",
-                                amount_cents: 1572000,
-                            },
-                            {
-                                currency: "NZD",
-                                amount_cents: 230000,
-                            },
-                            {
-                                currency: "JPY",
-                                amount_cents: 212590,
-                            },
-                            {
-                                currency: "BRL",
-                                amount_cents: 827000,
-                            },
-                            {
-                                currency: "BTC",
-                                amount_cents: 7537000,
-                            },
-                            {
-                                currency: "BCH",
-                                amount_cents: 1336700000,
-                            },
-                            {
-                                currency: "ETH",
-                                amount_cents: 1370270000,
-                            },
-                            {
-                                currency: "LTC",
-                                amount_cents: 2844800000,
-                            },
-                            {
-                                currency: "DOG",
-                                amount_cents: 1126500000000,
-                            },
-                            {
-                                currency: "USDT",
-                                amount_cents: 150000000000,
-                            },
-                            {
-                                currency: "BNB",
-                                amount_cents: 644380000,
-                            },
-                            {
-                                currency: "ADA",
-                                amount_cents: 6175000000,
-                            },
-                            {
-                                currency: "TRX",
-                                amount_cents: 29315000000,
-                            },
-                            {
-                                currency: "USDC",
-                                amount_cents: 1500000000,
-                            },
-                        ],
-                    },
-                    {
-                        field: "wager",
-                        value: 3,
-                    },
-                    {
-                        field: "max_win",
-                        value: [
-                            {
-                                currency: "EUR",
-                                amount_cents: 1000000,
-                            },
-                            {
-                                currency: "USD",
-                                amount_cents: 1000000,
-                            },
-                            {
-                                currency: "NOK",
-                                amount_cents: 10500000,
-                            },
-                            {
-                                currency: "CAD",
-                                amount_cents: 1500000,
-                            },
-                            {
-                                currency: "NZD",
-                                amount_cents: 1500000,
-                            },
-                            {
-                                currency: "JPY",
-                                amount_cents: 1416000,
-                            },
-                            {
-                                currency: "AUD",
-                                amount_cents: 1500000,
-                            },
-                            {
-                                currency: "BRL",
-                                amount_cents: 5500000,
-                            },
-                            {
-                                currency: "USDC",
-                                amount_cents: 10472000000,
-                            },
-                            {
-                                currency: "BTC",
-                                amount_cents: 2518000,
-                            },
-                            {
-                                currency: "ETH",
-                                amount_cents: 456780000,
-                            },
-                            {
-                                currency: "LTC",
-                                amount_cents: 947900000,
-                            },
-                            {
-                                currency: "BCH",
-                                amount_cents: 445400000,
-                            },
-                            {
-                                currency: "DOG",
-                                amount_cents: 375500000000,
-                            },
-                            {
-                                currency: "USDT",
-                                amount_cents: 1000000000000,
-                            },
-                            {
-                                currency: "BNB",
-                                amount_cents: 4295840000,
-                            },
-                            {
-                                currency: "ADA",
-                                amount_cents: 41164000000,
-                            },
-                            {
-                                currency: "TRX",
-                                amount_cents: 195433000000,
-                            },
-                        ],
-                    },
-                ],
-            },
-        ],
-        id: 767675,
-    },
-    {
-        weight: 5,
-        prize: false,
-        created_at: "2024-11-07T13:46:44.095Z",
-        bonuses: [
-            {
-                title: "Rocket Wheel Bonus 10 000 EUR",
-                type: "bonus",
-                attributes: [
-                    {
-                        field: "bonus_amount",
-                        value: [
-                            {
-                                currency: "EUR",
-                                amount_cents: 1000000,
-                            },
-                            {
-                                currency: "USD",
-                                amount_cents: 1000000,
-                            },
-                            {
-                                currency: "CAD",
-                                amount_cents: 1500000,
-                            },
-                            {
-                                currency: "AUD",
-                                amount_cents: 1500000,
-                            },
-                            {
-                                currency: "NOK",
-                                amount_cents: 10500000,
-                            },
-                            {
-                                currency: "NZD",
-                                amount_cents: 1500000,
-                            },
-                            {
-                                currency: "JPY",
-                                amount_cents: 1417145,
-                            },
-                            {
-                                currency: "BRL",
-                                amount_cents: 5512000,
-                            },
-                            {
-                                currency: "BTC",
-                                amount_cents: 50278000,
-                            },
-                            {
-                                currency: "BCH",
-                                amount_cents: 8894000000,
-                            },
-                            {
-                                currency: "ETH",
-                                amount_cents: 9138870000,
-                            },
-                            {
-                                currency: "LTC",
-                                amount_cents: 18983400000,
-                            },
-                            {
-                                currency: "DOG",
-                                amount_cents: 7510000000000,
-                            },
-                            {
-                                currency: "USDT",
-                                amount_cents: 1000000000000,
-                            },
-                            {
-                                currency: "BNB",
-                                amount_cents: 4297370000,
-                            },
-                            {
-                                currency: "ADA",
-                                amount_cents: 41151000000,
-                            },
-                            {
-                                currency: "TRX",
-                                amount_cents: 195371000000,
-                            },
-                            {
-                                currency: "USDC",
-                                amount_cents: 10000000000,
-                            },
-                        ],
-                    },
-                    {
-                        field: "wager",
-                        value: 3,
-                    },
-                    {
-                        field: "max_win",
-                        value: [
-                            {
-                                currency: "EUR",
-                                amount_cents: 3000000,
-                            },
-                            {
-                                currency: "USD",
-                                amount_cents: 3000000,
-                            },
-                            {
-                                currency: "NOK",
-                                amount_cents: 31438000,
-                            },
-                            {
-                                currency: "CAD",
-                                amount_cents: 4580000,
-                            },
-                            {
-                                currency: "NZD",
-                                amount_cents: 4580000,
-                            },
-                            {
-                                currency: "JPY",
-                                amount_cents: 4253235,
-                            },
-                            {
-                                currency: "AUD",
-                                amount_cents: 4580000,
-                            },
-                            {
-                                currency: "BRL",
-                                amount_cents: 16534500,
-                            },
-                            {
-                                currency: "USDC",
-                                amount_cents: 30000000000,
-                            },
-                            {
-                                currency: "BTC",
-                                amount_cents: 150833000,
-                            },
-                            {
-                                currency: "ETH",
-                                amount_cents: 27416620000,
-                            },
-                            {
-                                currency: "LTC",
-                                amount_cents: 56950200000,
-                            },
-                            {
-                                currency: "BCH",
-                                amount_cents: 26682000000,
-                            },
-                            {
-                                currency: "DOG",
-                                amount_cents: 22530000000000,
-                            },
-                            {
-                                currency: "USDT",
-                                amount_cents: 3000000000000,
-                            },
-                            {
-                                currency: "BNB",
-                                amount_cents: 12892100000,
-                            },
-                            {
-                                currency: "ADA",
-                                amount_cents: 123452000000,
-                            },
-                            {
-                                currency: "TRX",
-                                amount_cents: 586112000000,
-                            },
-                        ],
-                    },
-                ],
-            },
-        ],
-        id: 767676,
-    },
-];
