+++ src/store/ui.ts
@@ -3,7 +3,7 @@ import { defineStore } from "pinia";
 
 import { EventBus as bus } from "../plugins/EventBus";
 import bodyDisableScroll from "../helpers/bodyDisableScroll";
-import type { IModalOptions } from "../models/modalOptions";
+import type { IModalOptions } from "../types/modalOptions";
 
 type IFormInputsData = Record<string, string>;
 
@@ -15,22 +15,22 @@ export function createUIStore(theme: string) {
             login: "",
             password: "",
         });
-
+    
         const colorTheme = ref<string>(theme);
         const isThemeDark = ref<boolean>(colorTheme.value === "theme-dark");
-
+    
         function setShowModal(options: IModalOptions) {
             const modalIsOpen = modals.value.some((modal: IModalOptions) => {
                 return modal.name === options.name;
             });
-
+    
             if (!modalIsOpen) {
                 showModal.value = true;
                 modals.value.push(options);
                 bodyDisableScroll(true);
             }
         }
-
+    
         function closeModal({ name, immediate = false, ...args }) {
             const index = modals.value.findIndex((modal) => {
                 return modal.name === name;
@@ -38,7 +38,7 @@ export function createUIStore(theme: string) {
             if (index >= 0) {
                 modals.value.splice(index, 1);
             }
-
+    
             if (modals.value.length === 0) {
                 if (immediate) {
                     showModal.value = false;
@@ -53,27 +53,27 @@ export function createUIStore(theme: string) {
             }
             bus.$emit("modal.closed", name, { ...args });
         }
-
+    
         function setNewDataToFormInputs(fieldData: IFormInputsData) {
             formsInputs.value = {
                 ...formsInputs.value,
                 ...fieldData,
             };
         }
-
+    
         const getFormInputs = (key: string) => formsInputs.value[key];
-
+    
         return {
             colorTheme,
             modals,
             showModal,
-
+    
             isThemeDark,
-
+    
             setShowModal,
             closeModal,
             setNewDataToFormInputs,
-
+    
             getFormInputs,
         };
     });
