+++ src/types/configs.ts
